[core] Break up ddtrace.settings into sub-modules
