+++ ddtrace/settings/__init__.py
@@ -1,0 +1,13 @@
+from .config import Config
+from .exceptions import ConfigException
+from .http import HttpConfig
+from .hooks import Hooks
+from .integrations import IntegrationConfig
+
+__all__ = [
+    'Config',
+    'ConfigException',
+    'HttpConfig',
+    'Hooks',
+    'IntegrationConfig',
+]
